fix(client)!: backfill missing property descriptions on 12 schemas (test_schema_comprehensive) (#681)
The schema_validation pytest mark covers 12 schemas in docs/katana-openapi.yaml
whose properties were missing required descriptions, and one (Inventory) whose
default_storage_bin property had no type definition. These tests run under
`pytest -m schema_validation` (excluded from the canonical `poe test`) so the
regressions did not block PRs.

Fixes the following pre-existing failures:

Missing $ref-sibling descriptions (use-site role, per CLAUDE.md):
- CreatePurchaseOrderRequest.entity_type
- CreatePurchaseOrderAdditionalCostRowRequest.distribution_method
- UpdatePurchaseOrderAdditionalCostRowRequest.distribution_method
- CreateSalesOrderAddressRequest.entity_type
- CreateCustomerAddressRequest.entity_type

Missing inline property descriptions:
- UpdateCustomerAddressRequest.{first_name, last_name, company, phone, line_1,
  line_2, city, state, zip, country}
- CreateProductOperationRowItem.product_variant_id
- CreateProductOperationRowsRequest.rows
- LocationListResponse.data
- ProductOperationRowListResponse.data
- UnassignedBatchTransactionListResponse.data

Inventory.default_storage_bin: previously declared with only a description (no
type / $ref / allOf), generating as `Any | None` in both attrs and pydantic.
Typed as VariantDefaultStorageBinLinkResponse (the existing schema that matches
the wire shape: id, bin_name, variant_id, storage_bin_id, deleted_at). This
narrows the public type from `Any | None` to `VariantDefaultStorageBinLinkResponse | None`
on `Inventory.default_storage_bin` in both the attrs (`katana_public_api_client.models.inventory`)
and pydantic (`katana_public_api_client.models_pydantic._generated.inventory`)
surfaces.

CreateProductOperationRowItem.product_variant_id was also declared as
`type: number` (regenerating as `float`) — corrected to `type: integer` to
match the upstream gateway spec and the rest of the codebase's ID-typing
convention. This narrows the public type from `float` to `int` (with
`int` `minimum`/`maximum` constraints) on
`CreateProductOperationRowItem.product_variant_id` in both the attrs
(`katana_public_api_client.models.create_product_operation_row_item`) and
pydantic (`katana_public_api_client.models_pydantic._generated.inventory`)
surfaces. Caught by Copilot review on PR #681.

Regenerated via `uv run poe regenerate-client` and `uv run poe generate-pydantic`.
Generated-file impact: description fields added to inventory, contacts, common,
purchase_orders, sales_orders pydantic modules; Inventory attrs model gets
proper `VariantDefaultStorageBinLinkResponse` import and to_dict/from_dict
plumbing in place of the prior `Any` passthrough; CreateProductOperationRowItem
attrs + pydantic surfaces get `int` typing.

Bundles uv.lock drift from sibling-package release on main (mcp v0.69.0 → 0.70.0).

BREAKING CHANGE: Two narrowings in this PR.
- `Inventory.default_storage_bin`: `Any | None` → `VariantDefaultStorageBinLinkResponse | None`.
  Callers relying on the `Any` escape hatch must now treat the value as a
  `VariantDefaultStorageBinLinkResponse` (or `Unset` / `None`).
- `CreateProductOperationRowItem.product_variant_id`: `float` → `int`.
  Callers passing a `float` for this ID field must now pass an `int`. The
  field has always semantically been an integer ID; the float typing was
  a spec error, not a contract.

+++ katana_public_api_client/models/create_product_operation_row_item.py
@@ -15,7 +15,7 @@ T = TypeVar("T", bound="CreateProductOperationRowItem")
 class CreateProductOperationRowItem:
     """A single product operation row item in a bulk create request"""
 
-    product_variant_id: float
+    product_variant_id: int
     operation_id: int | Unset = UNSET
     operation_name: str | Unset = UNSET
     resource_id: int | Unset = UNSET
